Free

+++ structures_typedef/4-new_dog.c
@@ -16,7 +16,7 @@ return (NULL);
 
 new_dog->name = strdup(name);
 new_dog->owner = strdup(owner);
-new_dog->age = age;
+new_dog->age = (int)age;
 
 return (new_dog);
 free(new_dog->name);
